added a new function called domino_sum and  the user is prompted to enter the value of n
In the provided code, I added a new function called domino_sum, which calculates the sum of the first n terms in the domino sequence using both the domino and domino_fast functions. The results of this sum are printed in the if name == "main": block. and  the user is prompted to enter the value of n, and the code will calculate and display the results for the provided value of n.